[Refactor] #17: floor 패키지 속 class를 building 패키지로 이동
