+++ src/main/java/com/example/BarrierKU/domain/building/dto/SpaceSummary.java
@@ -1,4 +1,4 @@
-package com.example.BarrierKU.domain.floor.dto;
+package com.example.BarrierKU.domain.building.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
